Add card data parsing and sealed price fetching helpers; refactor imports in TCGScraper
