+++ Scraper/helpers/pull_rate_helper.py
